pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "algotrade"
version = "0.1.0"
description = "Algorithmic trading system with neural network predictions"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["trading", "algorithmic-trading", "machine-learning", "pytorch", "stocks"]

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "torch>=2.0.0",
    "yfinance>=0.2.0",
    "pandas-ta>=0.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["algotrade*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]